move the template functions to its own file
